+++ tests/UnitTests/Orderly.Domain.UnitTests/TestUtils/Quantity/QuantityAssertion.cs
@@ -5,8 +5,8 @@ namespace Orderly.Domain.UnitTests.TestUtils.Quantity;
 public sealed class QuantityAssertion : BaseAssertion
 {
     public static void AssertQuantity(
-        Domain.Quantity.ValueObjects.Quantity expected,
-        Domain.Quantity.ValueObjects.Quantity actual
+        Domain.Order.ValueObjects.Quantity expected,
+        Domain.Order.ValueObjects.Quantity actual
     )
     {
         Assert.NotNull(actual);
