+++ tests/UnitTests/Orderly.Domain.UnitTests/TestUtils/Quantity/QuantityFixture.cs
@@ -1,24 +1,24 @@
-using Orderly.Domain.Quantity.Validators;
+using Orderly.Domain.Order.Validators;
 
 namespace Orderly.Domain.UnitTests.TestUtils.Quantity;
 
 public sealed class QuantityFixture : BaseFixture
 {
-    private static Domain.Quantity.ValueObjects.Quantity CreateValidQuantity()
+    private static Domain.Order.ValueObjects.Quantity CreateValidQuantity()
     {
         var value = Faker.Random.Int(QuantityValidator.QuantityMin, QuantityValidator.QuantityMax);
         
-        return Domain.Quantity.ValueObjects.Quantity.Create(value);
+        return Domain.Order.ValueObjects.Quantity.Create(value);
     }
     
-    public static Domain.Quantity.ValueObjects.Quantity CreateQuantity(
-        Domain.Quantity.ValueObjects.Quantity? quantity = null,
+    public static Domain.Order.ValueObjects.Quantity CreateQuantity(
+        Domain.Order.ValueObjects.Quantity? quantity = null,
         int? value = null
     )
     {
         var lQuantity = quantity ?? CreateValidQuantity();
 
-        return Domain.Quantity.ValueObjects.Quantity.Create(value ?? lQuantity.Value);
+        return Domain.Order.ValueObjects.Quantity.Create(value ?? lQuantity.Value);
     }
     
     public static int CreateBellowMinQuantity()
